[][OpenWrt Dev][nmbm: add support to check the usability of a erased block]
[Description]
Add an option to specify the ability of the lower device which indicates
that whether the lower driver supports ECC correction on empty pages.

When this option is set, NMBM will read all pages of a just erased block.
If ECC error reported, the block is believed having too many bitflips and
should be marked bad.

[Release-log]
N/A

Change-Id: If837b97d4691ff6d41f11540ba1da295b2d45052
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/4649343