[][kernel][mt7988][pinctrl][separate sdcard from emmc45]
[Description]
Add sdcard pinctrl group to mediatek mt7988 SoC.

We need 11 pins for emmc 4.5 and 12 pins for emmc 5.4, but if we just
use emmc as SD card controller, we only need MSDC0_CLK, MSDC0_CMD and
MSDC0_DATA0/1/2/3 which belong the subset of emmc 4.5 pins.

This patch add a new pinctrl group 'sdcard' to use fewer pins to make
sdcard works.

[Release-log]
N/A

Change-Id: Idc3e80929111abe6d8ddf384fb4894253359dd66
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/6722624