[][MAC80211][hostapd][Add the broadcast address of Probe Response frame on 6 GHz]
[Description]
Add the broadcast destination address of Probe Response frame on 6 GHz.

According to IEEE 802.11ax 26.17.2.3.2,
if a 6 GHz AP receives a Probe Request frame and responds with
a Probe Response frame, the Address 1 field of the Probe Response
frame shall be set to the broadcast address,
unless the AP is not indicating its actual SSID inthe SSID element
of its Beacon frames.

Without this, hostapd fill the Address 1 feild of the Probe Response
frame with the source address from Probe Request frame on 6 GHz.
Fix this to use broadcast address instead.

[Release-log]
N/A

Change-Id: I12e30f8fd179f20037879dffe24c0118636d0a89
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7057107