[][mac80211][mt76][move temperature margin check to mt7915_thermal_temp_store()]
[Description]
Change the place to do the temperature margin check.

Originally, we would reduce the 10-degree margin to the restore
temperature, but the user would not be aware of this when setting it.
Moving the margin reduction to the user setting check allows the user to
clearly understand that there is a 10-degree difference between the
restore and trigger temperature.

[Release-log]
N/A

Change-Id: I81d14eb441229d32afb497177d750a1995a56b3a
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7744133