[][MAC80211][hnat][Add UL/DL HQoS support]
[Description]
Add UL/DL HQoS support.

We add a new design to the HQoS, separating CONNMARK into low and
high words.
 - The bits 0 to 15 of CONNMARK are utilized to assign the TX Queue
   for WAN to LAN/WLAN packets.
 - The bits 16 to 31 of CONNMARK are utilized to assign the TX Queue
   for LAN/WLAN to WAN packets.
To activate this feature, the user needs to add the TX Queue index
to the high word of the CONNMARK.

Without this patch, the PPE driver is unable to allocate UL/DL
packets to their corresponding TX Queue.

[Release-log]
N/A


Change-Id: I3bd8916bfd7286ff4352b0bbdbbc3838eb79c97d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7747036