[][kernel][mt7981][eth][Enable Maxlinear GPHY driver]
[Description]
Change Linux configuration to enable Maxlinear GPHY driver.

In the previous commit with ID 7248826, we updated the attach
PHY driver method to use probe ID instead.
However, we did not to enable the Maxlinear GPHY driver.

Without this patch, the Maxlinear GPY211 will not be able to work
normally.

[Release-log]
N/A


Change-Id: I8b4a213c5ef9208a292c218bece1a6c04f8ae460
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7806602