[][openwrt][mt7988][tops][tnl-offload: support with crypto-eip]
[Description]
Add tunnel offload with crypto-eip support.

Tunnel parameter now records CDRT index to provide help for tunnel
decapsulation and encapsulation offload setup flow.

For decapsulation, If skb already carried with CDRT information, it means
that packet is already decrypted by HW offload engine crypto-eip and there
is already a CLS rule setup for that packet. Therefore, we should not
modify the exist CLS rule. Instead, we just need to update the TPORT to
that CLS entry.

For encapsulation, there is nothing more need to do.

[Release-log]
N/A

Change-Id: Ib4ff408836fa0ed16895fc486c5ebdee90cfa0b6
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7923905