[][MAC80211][hnat][Update 2.5G LAN/WAN low t-Put issue for the PPPQ]
[Description]
Refactor 2.5G LAN/WAN low t-Put patch for the PPPQ.

In the previous commit with ID 7905304, we allocated WLAN to 2.5G LAN
packets to QDMA TX Queue5 in the PPPQ mode.
However, we discovered that the packets do not pass through Queue5
due to missing sp assignment in the mtk_ppe_check_pppq_path().
It led to the poor performance in the WLAN to 2.5G LAN scenario.

Without this patch, the user may experience the poor perforamce in
the WLAN to 2.5G LAN scenario.

[Release-log]
N/A


Change-Id: I85a9da200ef5d732cfaa88d648f2123d0682fac7
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7932086