[][openwrt][common][eth][Add a Airoha EN8811H 2.5G PHY driver package]
[Description]
Add a Airoha EN8811H 2.5G PHY driver package.

Airoha recommends using a kernel module to load the EN8811H PHY driver.
If the user does not utilize a kernel module to load the PHY driver,
the PHY driver may have failed to download the firmware.

Without this patch, the PHY driver may not download the firmware
successfully.

[Release-log]
N/A


Change-Id: I8694aba81cd0ed6fad138bbf877568d281cb7d92
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/7980789