[][MAC80211][hnat][Fix PPE hanging issue during the reboot test]
[Description]
Fix PPE hanging issue during the reboot test.

In the mtk_ppe_stop() function, the PPE scan mode is not disabled before
disabling the PPE. This can potentially lead to a hang during the process
of disabling the PPE.

Without this patch, the PPE may experience a hang during the reboot test.

[Release-log]
N/A


Change-Id: I3bc692c3e43f06df6d7f5c0205e45870ce029eee
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8140700