[][MAC80211][WiFi6/7][app][Fix atenl sync eeprom all issue]
[Description]
Fix atenl sync eeprom all issue
If two or more chip devices use flash mode, it is possibe to cross
the user-specified boundary when we call "sync eeprom all" to
write the entire atenl eeprom temp file into mtd.
Note that the size of atenl eeprom temp file is fixed
at 0xFF000 for Eagle pre-cal.
Therefore, we utilize the dd cmd to slice the file to
a proper size before writting it to mtd.

[Release-log]
N/A

Change-Id: Id26d757d5422ef985b5775b0c2c1daa007905676
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8234404