[][kernel][common][eth][Fix issue with XGDM hang when reconnecting SFP+ Optical module]
[Description]
Fix issue with XGDM hang when reconnecting SFP+ Optical module.

This patch corrects the down/up sequence of the XGAMC link and adds a
delay between the USXGMII PCS link up and the XGMAC link up to avoid
runt packet issues.

Without this patch, the users may encounter a traffic stuck issue when
reconnecting the SFP+ optical module.

[Release-log]
N/A


Change-Id: I3981638e9be270ae8256cdc4070c319c6cc80e8b
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8269169