[][kernel][mt7988][eth][Update panic issue during SER test on the 4GB RFB]
[Description]
Refactor panic issue during SER test on the 4GB RFB.

If the mtk_poll_rx() function detects the MTK_RESETTING flag, it will
jump to release_desc and refill the high word of the SDP on the 4GB RFB.
Afterwards, when mtk_rx_clean processes an incorrect SDP, it can lead
to a panic. However, the previous change does not clear the DONE bit of
RXD2, which has the potential to cause an ADMA hang in this abnormal
descriptor.

Without this patch, the kernel will panic in the mtk_rx_clean() function
during SER test.

[Release-log]
N/A


Change-Id: I9cfb59a4bc8b5b3536b50e6b7476bc50ac1892e8
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8456306