[][kernel][common][hnat][Increase the field width of the SW_UDF]
[Description]
Change the field width of the SW_UDF.

In the previous change, we used 1 bit of the SW_UDF to implement an
entry lock mechanism for the WiFi TX scenario. However, this change
may result in customer applications not functioning properly.
As a solution, we increase the field width of the SW_UDF in order
to meet customer requirements.

Without this patch, the customer applications not functioning
properly.

[Release-log]
N/A


Change-Id: I9608302b2b0ce8ff0c98c8e098da389671f81d9e
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8468033