[][HIGH][kernel][common][eth][Fix HQoS is not functioning properly following the SER issue]
[Description]
Fix HQoS is not functioning properly following the SER issue.

Since SER will reset QDMA, it means that the HQoS configurations will
be cleared. The driver needs to save the HQoS configuration before the
SER and then restore the HQoS configuration afterward.

Without this patch, the HQoS is not functioning properly following the
SER.

[Release-log]
N/A


Change-Id: Ibbdc8476bc7081ffcc34bdc7a92babeb5891e6f2
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9095074