[][HIGH][kernel][common][eth][Fix the GDM RX FC packet statistics error issue]
[Description]
Fix the GDM RX FC packet statistics error issue.

The GDM RX FC hardware counter is designed to clear data
after reading, so the statistical values are reset to
zero once they are read.

We have found that both the GDM RX FC packet statistics
function and the GDM monitoring function access the
hardware counter, which leads to data contention and
results in incorrect statistical values. There must be
only one single place to read the hardware counter.

Without this patch, the GDM RX FC packet count statistics
in Ethernet driver will be incorrect, and it may lead to
the GDM monitor misjudging the status of QDMA.

[Release-log]
N/A


Change-Id: I320c6c11a920b47ff432f8e8794cbfa2c4dee18e
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9411054