[][kernel][mt7988][eth][Refactor error detection rule for the WDMA_RX monitor]
[Description]
Refactor error detection rule for the WDMA_RX monitor.

In this update, we are introducing a new rule to identify when CONNSYS
is busy if the WDMA RX ring has packets that need to be processed, but
the CIDX and DIDX of WDMA RX remain unchanged.
Additionally, we are refactoring the existing rule to identify when
NETSYS is busy if there are packets in the PSE to WDMA output queue
waiting to be written into the WDMA RX ring, but the CIDX and DIDX of
WDMA RX remain unchanged.

Without this patch, certain error conditions will not be identified in
the WDMA_RX monitor.

[Release-log]
N/A


Change-Id: I5f142bc9303c5c6eabaca6b872ece4a8c5d4295b
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9505871