[][openwrt][mt7988][crypto][Add timer for DDK interrupt]
[Description]
Add timer for DDK interrupt.
In DDK, we use PEC_ResultNotify_Request() to invoke packet-coalescing
interrupts. However, there are no time-coalescing in EIP. If the packet
number for invoking interrupt is too small, performance is not good. Too
high then connection with only few packets will disconnected. So, we add
a timer to make a balance between them.

[Release-log]
N/A


Change-Id: I3513b33dea454289b5b3f04b19aa51112c51501d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9516358