[][kernel][common][hnat][Add HNAT DSA tag for MXL862_8021Q protocol]
[Description]
Add HNAT DSA tag for MXL862_8021Q protocol.

For the switch of MXL862_8021Q protocol, using VLAN tag rather than
special tag for HNAT offload, in this patch, we apply different DSA tag
based on the DSA protocol of the device to support the DSA device with
MXL862_8021Q protocol.

[Release-log]
N/A


Change-Id: Ib776ac463b9d710e515f022c320c5d25d3351421
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9589132