[][kernel][common][eth][Fix the ethX transmit timeout issue caused by the NETSYS SER]
[Description]
Fix the ethX transmit timeout issue caused by the NETSYS SER.

In the previous change 9787474, we added a new NETSYS SER design to
prevent resetting GMAC when GMAC is not hung. However, this design
introduced a 'ethX: transmit timed out' issue when the Ethernet port
is not connected to the link partner.
Therefore, we added a check to avoid mistakenly calling
netif_carrier_on().

Without this patch, the kernel will encounter an 'ethX: transmit timed
out' issue after NETSYS SER.

[Release-log]
N/A


Change-Id: I5360d4fa985a2890373c438c8938c50ba05ae0bf
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9821872