[][kernel][common][eth][Refactor NETSYS SER to prevent resetting GMAC when GMAC is not hang]
[Description]
Refactor NETSYS SER to prevent resetting GMAC when GMAC is not hang.

This new design introduces a potential risk: it does not reset the GMAC
during the booting procedure, which may cause the GMAC to hang during
a reboot in traffic flooding tests.

Without this patch, the GMAC may experience a hang issue that caused
by a system reboot during the traffic flooding test.

[Release-log]
N/A


Change-Id: If98bb36e6b14f1374a38c74ba0837e64341f3e8a
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9927140