[][kernel][common][eth][Add PSE port6 link down before disabling QDMA Tx for the NETSYS SER]
[Description]
Add PSE port6 link down before disabling QDMA Tx for the NETSYS SER.

When the hardware path routes to the QDMA, it will queue at the output
queue of PSE_P6 until QDMA transmits packets to the FP. Therefore, we
need to force link down PSE_P6 and wait for the QDMA transmits all the
packets before disabling QDMA Tx.

Without this patch, the NETSYS warm reset may fail because the FE does
not return to an idle state.

[Release-log]
N/A


Change-Id: I5b2450c1813fbecf1e6b07842615998ec28c271c
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9984161