[][kernel][mt7987][eth][Add SGMII reconfiguration to the mtk_mac_fe_reset_complete() for NETSYS SER fast mode]
[Description]
Add SGMII reconfiguration to the mtk_mac_fe_reset_complete() for NETSYS
SER fast mode.

In the current flow, the NETSYS SER fast mode powers down the SGMII
in mtk_stop(), but it does not to reconfigure the SGMII. Therefore,
the SGMII may not function properly after releasing power down the
SGMII in mtk_open().

Without this patch, the SGMII may not function properly after triggering
the NETSYS SER fast mode.

[Release-log]
N/A


Change-Id: Icc5d5a71df15099268278418d809ab75b363ddd1
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/9984162