[][kernel-6.6][common][eth][Add MIB read feature to the Aquantia PHY driver]
[Description]
Add MIB read feature to the Aquantia PHY driver.

The users can utilize following command to dump the MIB information
from AQR113C.
  - cat /sys/kernel/debug/aquantia/phyX; X is 0 or 8

This feature is available only on the link partners with speeds of
2.5G, 5G, and 10G.

Without this patch, it is difficult to locate the source of packet
loss, whether it is on the system side or the line side.

[Release-log]
N/A


Change-Id: If0f179881607799dc3a23520e4dc10403e08ea7e
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10000503