[][package][network][ethtool][Change ethtool to version 6.7]
[Description]
Change ethtool to version 6.7.

Currently, OpenWrt 24.10 + kernel 6.6 is using ethtool version 6.10.
However, we have discovered that ethtool 6.10 has a structure that
does not match with the kernel.

We will first ensure that the functionality of ethtool can be used
normally by downgrading the version. In the future, we will consider
whether to add a kernel patch.

[Release-log]
N/A


Change-Id: I02112202b3b8ab149c397c1ef5d011c3e42fd36a
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10182944