[][kernel][common][hnat][Fix PPE index mapping issue in PPE static mode]
[Description]
Fix PPE index mapping issue in PPE static mode.

Without this patch, the PPE static entry would be created in
PPE0 when CFG_PPE_NUM == 2 and sport == GMAC2, but the
packets from GMAC2 is sent to PPE1 when CFG_PPE_NUM >= 2,
so the static PPE entry would not be hit by GMAC2 packets.

[Release-log]
N/A


Change-Id: I018c7ca0bbc15ea1c49d53efa5a05bd94e9666c3
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10190754