[][kernel][mt7988/87][thermal][Add thermal throttling support]
[Description]
Add thermal throttling support with dynamic CPU frequency scaling.

We use the Linux kernel CPU cooling framework to dynamically
adjust the CPU frequency and Vproc voltage based on the Tj
temperature. The CPU frequency is set according to the OPP table.

The kernel enables the step-wise governor by default. When the Tj
temperature is higher than or lower than the trip point, the CPU
frequency will be adjusted in steps. There will be no conflict
when the step-wise governor and the ondemand governor are
operating simultaneously. When the Tj temperature is higher than
the trip point, the settings of the step-wise governor will take
precedence. When the Tj temperature is lower than the trip point,
the settings of the ondemand governor will take precedence.

By default, we set the trip point for frequency scaling at 120°C
to prevent the IC from triggering a thermal reset.

Without this patch, there is no proactive cooling behavior before
triggering a thermal reset.

[Release-log]
N/A


Change-Id: Ifc67a068207f3d91ae5d7e1a0d7493713e7a41ed
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10255756