[][kernel][mt7987][eth][Fix Tx timestamp abnormal issue with the IEEE1588v2 during GMAC congestion]
[Description]
Fix Tx timestamp abnormal issue with the IEEE1588v2 during GMAC
congestion.

In the original process, the PTP driver reports the hardware timestamp
to ptp4l when QDMA transmit PTP packet successfully. However, at that
time, if there is network congestion, the GMAC may not transmit this
packet to the network. Consequently, the hardware timestamp will remain
unchanged. Ultimately, this leads to the PTP driver reporting an
incorrect hardware Tx timestamp back to ptp4l.
To resolve this error scenario, we introduce a PTP tx work to wait for
GMAC Tx completion, which avoids blocking CPU in the napi_tx process.

Without this patch, the ptp4l might receive an incorrect hardware Tx
timestamp.

[Release-log]
N/A


Change-Id: Ibdcbcb38ab6e7f1e2c5ee5a78393fd72d653daf0
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10259222