[][kernel][common][hnat][Fix issue of iptables rule no effect on PPE entry when bridge-nf-call-iptables is enabled]
[Description]
Fix issue of iptables rule no effect on PPE entry when
bridge-nf-call-iptables is enabled.

Without this patch, if the iptables rule is inserted in
post-routing chain, because our original flow is to set PPE BIND
entry in bridge post-routing hook for bridge layer traffic,
which is before the IP-layer post-routing hook so the iptables rule
would not have effect on PPE entry of bridge layer traffic.

[Release-log]
N/A


Change-Id: I9a7b8956b08b77f8b76aef770f95eabc89994008
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10281158