[][kernel-6.6][common][switch][Change agetime setting method for DSA switch]
[Description]
Change agetime setting method for DSA switch.
The agetime field of the switch MAC table will be reset to the register
AAC.AGE_CNT and counted down by one every AAC.AGE_UNIT seconds.
To make it more readable, overwrite the upstream method by forcing
AAC.AGE_UNIT to 1 so it counts down by one second,
and the setting range of agetime will be 1 to 512 because AGE_CNT_MAX
is 0xff.

[Release-log]
N/A


Change-Id: Iee0740fd7d1662638f3da2ef80c3d13e13fdaa0d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10370125