[][openwrt-24][MAC80211][hnat][Improve ETH to ETH performace in an unbalanced PHY rate test for the mt7988]
[Description]
Add a patch to dispatch short packets to a high-priority queue in the
PPPQ path to improve ETH to ETH performance during an unbalanced PHY
rate test for the MT7988.

Without this patch, the performance of ETH to ETH cannot reach line rate
in an unbalanced PHY rate test for the mt7988.

[Release-log]
N/A


Change-Id: I420c1e0493f5f453eea3f0c0d7c32c232c5d2d31
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10380409