[][openwrt-24][MAC80211][hnat][Refactor DSCP learning flow]
[Description]
Refactor DSCP learning flow.

In the previous method, we learned the DSCP value from an SKB when it
passed through the flowoffload_tg(). Therefore, this method was unable
to learn the DSCP value from bi-directional SKBs. To address this
limitation, we implemented a new mechanism to learn the DSCP until count
of bi-directional TCP packets reached 30.
If a user would like to enable this feature, they can use the command
below to activate it.
 - echo 1 > /proc/sys/net/netfilter/nf_conntrack_qos

Without this patch, the PPEs are unable to learn the bi-directional DSCP
from SKBs.

[Release-log]
N/A


Change-Id: Ib51c8a92bd254429bc5be63379517c22b79065fa
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10395152