[][kernel-6.6][mt7988][secure-boot][add patch for parsing and mounting ramdisk as fitblk]
[Description]
Add patch for parsing and mount ramdisk as fitblk. For eMMC device,
rootfs_data is splited from firmware partition and mount as fitrw. When
using ramdisk as rootfs, we need to parse and mount the ramdisk node as
fitblk, then mount the remaining space as fitrw for rootfs_data.

[Release-log]
N/A

Change-Id: Id38624122cac563d35a0f83cdc0d00bbfd224f6d
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10510790