[][mac80211][wifi6][mt76][fix sanity check in mt7915_mac_tx_free]
[Description]
Fix sanity check in mt7915_mac_tx_free.
In some saces, the wlan id in txfree would be zero but the reset of
token id is valid and should be free. Free the reset of token id when
wlan id is zero or it would lead to token leakage.

[Release-log]
N/A

Change-Id: I793dd77d38f46f433a104f58d58a72eebebb3ed0
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10516102