[][kernel-6.6][common][eth][Change TX/RX NAPI poll weight]
[Description]
Change TX/RX NAPI poll weight.

The default parameter of 64 can impact the SW path peak throughput
by up to approximately 1.2Gbps. Therefore, we have adjusted the
napi poll weight parameter in the driver to 256 to align with the
configuration of kernel-5.4.

Without this patch, the SW path throughput will be significantly
affected.

[Release-log]
N/A


Change-Id: Ib847f58c1babb0f41046188a4c88ce53b3c5dcaf
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10593686