[][kernel-6.6][mt7988][eth][Change to software reset for the Aquantia 10G PHY]
[Description]
Change to software reset for the Aquantia 10G PHY.

In the current hardware design, the PHY will be reset by the GPIO when
bringing up the interface, and then the PHY driver will initiate the
firmware loading procedure again. However, this operation may take a
few seconds and cause WiFi disconnection during NETSYS SER.
Therefore, we utilize a software reset when starting up the system
to instead of using the MDIO GPIO reset.

Without this patch, the WiFi might disconnect during NETSYS SER.

[Release-log]
N/A


Change-Id: I93e5950fcd3025f96e217728ad85d052ba723be7
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10606760