[][openwrt-21][common][musl-upgrade][package: network: service: upgrade ppp to 2.4.9]
[Description]
Update the ppp package to version 2.4.9. The previous version of the
ppp package's header file (pppd.h) is used by feeds/packages/net/pptpd,
and this header will fail to build with the newer musl libc. To prevent
build failures in feeds/packages/net/pptpd with the newer musl libc,
we should also upgrade the ppp package to a newer version.

[Release-log]
N/A

Change-Id: Ifce5c840424c25435973dba0bea5a742a687806a
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10700537