[][openwrt][mt7988][usb][fix USB port0 function and enable U2 in 4 PCIe case]
[Description]
Fix the functionality of USB port0 U2 when U2 is enabled without U3.
This change addresses the issue where port0 U3 is shared with PCIE2,
ensuring that the port0 U2 function operates correctly without U3 support.

Additionally, add support to enable the U2 function instead of disabling
the entire USB port0 in the configuration for the 4 PCIe case. This
change ensures that U2 functionality is properly activated.

[Release-log]
N/A

Change-Id: I5dcd13d38767b26c00ef8f4d6c7f0b9aa6348f08
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10728298