[][openwrt-24][mt7988][eth][Refactor the issue of firmware loss after toggling the interface up and down for the Aquantia 10G PHY]
[Description]
Refactor the issue of firmware loss after toggling the interface up and
down for the Aquantia 10G PHY.

Some customers wish to keep the firmware load operation in the PHY probe
stage, so we have added it back. Simply add the firmware load operation
to the config_init function to ensure that the firmware can be restored
after it is cleared by the GPIO reset.

[Release-log]
N/A


Change-Id: I391ea187d42a3becd72c2450303cbc0379cc7efb
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10755744