[][openwrt-24][MAC80211][hnat][Fix issue of falsely delete entries in the WiFi Roaming feature]
[Description]
Fix issue of falsely delete entries in the WiFi Roaming feature.

In the AP and APCLI test cases, we found that the Roaming handler might
receive RTM_NEWNEIGH events from ETH devices, which could lead to the
deletion of PPE entries. As a result, users may observe a performance
drop in ixChariot. To address this issue, we need to add an additional
condition to filter out such events from ETH devices.

Without this patch, users may observe a performance drop in ixChariot
during AP and APCLI testing.

[Release-log]
N/A


Change-Id: If1162365089e258b32a290abd7e598de2b26a478
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10764032