[][openwrt][mt7988][hnat][Change the key for deleting crypto PPE entry]
[Description]
Change the key for deleting crypto PPE entry.
For IPSec inline mode, when rekeying, we need to delete PPE entry and
let packet go to CPU for PPE re-learning.
In the past, we use srouce IP and dest. IP from xfrm policy as the key
to search PPE entry and delete it. However, this only work on pure
IPSec and is not compatible with other scenario. Such as L2TP/IPSec, the
IP address on PPE is not same as IP address in xfrm policy.
This patch change the key for deleting crypto PPE entry from IP address
to cdrt_idx, which is more suitable and general for all use cases.

[Release-log]
N/A


Change-Id: If4406c9b604a2bf0092bfbf6e38785e0fff3ed89
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10830908