[][openwrt][MAC80211][hnat][Refactor DSCP learning flow for the xt_FLOWOFFLOAD]
[Description]
Refactor DSCP learning flow.

In the previous method, we learned the DSCP value from an SKB when it
passed through the flowoffload_tg(). Therefore, this method was unable
to learn the DSCP value from bi-directional SKBs. To address this
limitation, we implemented a new mechanism to learn the DSCP until count
of bi-directional TCP packets reached 30.
If a user would like to disable this feature, they can use the command
below to activate it.
 - echo 0 > /proc/sys/net/netfilter/nf_conntrack_qos

Without this patch, the PPEs are unable to learn the bi-directional DSCP
from SKBs.

[Release-log]
N/A


Change-Id: I3f71b7add78befb3567d34a712e45aabbb41d3e3
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10873246