[][openwrt][MAC80211][hnat][Fix issue of falsely delete entries in the WiFi Roaming feature]
[Description]
Fix issue of falsely delete entries in the WiFi Roaming feature.

In the AP and APCLI test cases, we found that the Roaming handler might
receive RTM_NEWNEIGH events from ETH devices, which could lead to the
deletion of PPE entries. As a result, users may observe a performance
drop in ixChariot. To address this issue, we need to add an additional
condition to filter out such events from ETH devices.

Without this patch, users may observe a performance drop in ixChariot
during AP and APCLI testing.

[Release-log]
N/A


Change-Id: Iba64129385b13930f8a4b6121c1b832f226db40f
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10899971