[][openwrt-24][[MAC80211][hnat][Add Keep DSCP feature for NETSYSv3]
[Description]
Add Keep DSCP feature for NETSYSv3.

Some use scenarios require preserving the DSCP value when packets are
forwarded by the PPE. Therefore, we have designed a new mode to achieve
this. Users can enable this feature using the following commands.
 - echo 1 > /sys/kernel/debug/mtk_ppe/dscp_toggle

Without this patch, the DSCP value may be changed if the DSCP field in
IB2 is not the same as that in the ingress packets.

[Release-log]
N/A


Change-Id: I48605111ea00615996fc0315dfbbc3a4aff66fae
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10950263