[][openwrt-24][MAC80211][eth][Fix IPSec traffic not entering SW fast path issue]
[Description]
Fix IPSec traffic not entering SW fast path issue.

Without this patch, the IPSec traffic conntrack would not be set to
OFFLOAD due to dscp learning function could not count the encry
direction's packet count to qos counter, then the TCP Bi-Di > 30
packets condition could not be fulfilled, leads to the OFFLOAD
rule could not be set.

[Release-log]
N/A


Change-Id: I446d7798de4faeaeb472af2da09cb6ea8147b2e5
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10951475