[][openwrt][MAC80211][hnat][Refactor the DSCP learning flow to include support for the SW fastpath]
[Description]
Refactor the DSCP learning flow to include support for the SW fastpath.

In the previous change, we added a DSCP learning feature to the
flowtable framework; however, it only passes the DSCP learning
result to the HW offload and does not update DSCP in SW fastpath.

Without this patch, egress traffic may not carry the correct DSCP
value when hitting the flowtable rule in the SW fastpath.

[Release-log]
N/A


Change-Id: I08483d76b8448b6324523e13996b4272d10389fd
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11038121