[][openwrt][MAC80211][hnat][Add dynamic control for DSCP learning threshold]
[Description]
Add dynamic control for DSCP learning threshold.

We have changed the default DSCP learning threshold from 30 packets
to 3 packets and added a dynamic control feature that allows users to
modify this threshold. As a result, users can now adjust the dynamic
learning threshold within the range of 0 to 200 packets, if necessary.
 - echo [0, 200] > /proc/sys/net/netfilter/nf_conntrack_qos

[Release-log]
N/A


Change-Id: I9652ede3a30dab0aa06bd97b44cf998ed544b8a6
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/11051653